+++ src/se/liu/ida/groupl2/tddd78/projekt/Player.java
@@ -14,8 +14,8 @@ import java.awt.Color;
 public class Player implements Collidable
 {
 
-    final static int PLAYER_SIZE = 35;
-    final static int MOVE_STEP = 5;
+    final static double PLAYER_SIZE = 35;
+    final static double MOVE_STEP = 5;
     final static int MAX_HP = 100;
     final static Color COLOR = Color.black;
     private int health;
